Refactoring controller logic
